ocr-web-app/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3

# OCR Dependencies
PyMuPDF==1.23.8
Pillow==10.2.0
pytesseract==0.3.10
opencv-python-headless==4.9.0.80
numpy==1.26.3
google-generativeai==0.3.2

